Create rcoin.h
Rcoin - A coin of Respect, by namzezam

Rcoin is a time limited and equally re-distributed coin - a cyclic and communal
coin in Community Exchange Systems (ces) - rings. Such rings are optionally
owned equally as a Common companies (ComCom) for becoming more open in
the bottom-up rebuild of people's economy. Use Rcoins to add social justice and
resilience to your community and yourself.

The Rcoin is daily and gradually losing its value, which is equally gained by 
the community members. As such, it is not money, nor an equivalent to money, but
still a medium of exchange or credit and an inner community evaluation tool.
  Like any thing besides money, Rcoin is going to die and be dispersed - 
    it's a natural medium. Its lifetime is adjustable per each cycle of 
    issuing new coins.
  Its lost value, as equivalent to percentage of inflation, is hundred 
    divided by its lifetime. Example: 50 years lifetime is like 2% inflation.
  This lost value is only lost for its owner, since the lost is gained as an 
    equal dividend by the community members. Hence, saving in isolation, alone 
    and not communal, now costs more to the owner too and not only to the 
    community.
  As a communal tax, the lost value taxes the saving-in-isolation for turning a
     toxic-competitive environments to become more supportive, wherein
     community’s members are benefited from advancing those who earn more,
     now also for the benefit of each member in that community.
     
The Rcoin is implemented as an electronic p2p money too, for making it easier
to automate its features. Its exchange In relation to the others can be direct 
or a substitutable one, meaning through the price of goods, services and 
holdings. It's the next after Bitcoin being the p2p's gold. Since it is only 
1) used in the ring with 2) constant lost value, which is 3) equally 
redistributed, trading in it makes it harder to prov “predictability” even 
after passing measures of “negligibility”.

Yes, people want immortal property, as they fear death and discriminate for
serving their debt in slavery until their death. Yes, people admire leaders 
doing things they are not legitimated to do. And yes - the community they are in
must have growth, while they prepare wars. Wars prepared to erase economies to 
be restarted again and again. Wars prepared also by exporting weapons,
corrupting criminal governments and exploiting others' land.
        But what is the medium forcing people to become such? Isn't it money and
mutual or speculative trust in immortal, accumulative and inheritable property?

        Systems allowing endless accumulation of power can only turn evil before
        collapsing. The medium makes the system and it's the system, stupid. 
                        Let's decentralize it!

For your community and yourself benefit, against gentrification and for blocking
fat fingers from taking over you or your cause, use money only to buy things
from far by your ring, and Rcoins for buying thing from near in your ring.

ComCom – A Common Company

A Common company (ComCom) is based on unchangeable agreement of
ownership between its owners matching the group's criteria and categorized
either as peers (equal-owners) or not, where
● each peer may only be a human-person or a ComCom,
 ● but a non-peer may be a human-person, ComCom or a non-ComCom
body;
● through all holding layers only a single holding position per owner is allowed;
● the group has the first right to buy any ownership over it
 ● and once in its creation it can become sealed allowing no more owners
over it;
● (d) is a positive number less than one representing the group's
decentralization,
 ● where all the peers together equally hold (d) times 100 percent of the
group;
● (c) is the count of the peers - to be decided independently by them alone;
● (v) is the ownership value of one peer - it worth peer-coin in Rcoin or money
and it ranges only proportionally to constant number times peer's ratio in group;
● (m) is the group's market value; and
● (v) reflects (m), such that (m) times (d) equals (c) times (v), where either 
(d) is constant, which is to be decided only once in the comcom's creation, or
 ● all owners of the group or of its sub groups are its members, of which
number is the same as of group's shares, each member can always return
to, and must first become, the group's peer and the number of shares held
by one peer is a positive number less than one, which is decided
independently only by the group's peers.

Here is a simple example of creating a comcomized network of peers, as the
ownership over some trademark of the network is shared/represented-by the
peers and an external association functions as a caretaker of the comcom, so
that such association could be swapped periodically or in situations of
dissatisfaction, while the association arranges personal agreements between
the members, such that
        the newcomer accept the agreement constituting the comcom,
        the newcomer commits that acceptance to each of the comcom's members
          by signing on attached list of all those members and
        each of the comcom's members commits the acceptance of that agreement
          to the newcomer by being represented by the association.
                
                For more try http://yes-i-am-angry.wikidot.com
                
Let's get Free and Equal people owning as teams ComComs
and trading as individuals in Rcoin: http://is-with.wikidot.com